+++ libs/rxjs/boolean/src/lib/is-truthy.ts
@@ -1,0 +1,17 @@
+/**
+ * @packageDocumentation
+ * @module Boolean
+ */
+import { OperatorFunction } from 'rxjs';
+import { map } from 'rxjs/operators';
+
+/**
+ * Returns an Observable boolean for whether each source value is truthy (`Boolean(value)`).
+ *
+ * @category Query
+ *
+ * @returns Observable that emits whether the value is truthy
+ */
+export function isTruthy(): OperatorFunction<unknown, boolean> {
+  return (source) => source.pipe(map((value) => Boolean(value)));
+}